Add files via upload
This database shows the daily traffic by Metrobús CDMX line. The data covers the period from July 2005. This database is updated monthly.

Broken down influx refers to the influx in the Public Transport Organization, separated by means of access, which are: Integrated Single Mobility Card and free access, the latter refers to free access for senior citizens, people with disabilities and/or children under 5 years of age, according to the requirements established by the Organization.

Updates after January 2022:
As a result of the 2022 Opening Plan , where a greater breakdown and granularity of the influx databases was required, updates from this date onwards will be reflected only in the Broken Down Influx database . From which, the total influx can be obtained from the afluencia total column.