Add files via upload
As a result of the physical distancing measures adopted in response to the coronavirus pandemic, there was a temporary reduction in the daily mobility of Mexico City residents. Beginning in the second half of March 2020, significant decreases were observed in the use of public and private transportation.

In this context, daily mobility monitoring was carried out to monitor the behavior of the population during the pandemic and make health-related decisions. Among the data sets collected for mobility monitoring are the traffic indices of Mexico City's transportation agencies (Metrobús).